requirements-gpu.txt
astor
gast
google-pasta
grpcio
h5py
Keras-Applications
Keras-Preprocessing
Markdown
numpy
opencv-python
Pillow
protobuf==3.20
PySide6
shiboken6
scipy
six
tensorboard
tensorflow-estimator
termcolor
Werkzeug
wrapt
scikit-image
imageio
pywavelets
